Добавить SwiftLint. Поправить замечания
