+++ .gitignore
@@ -1,1 +1,59 @@
 .idea
+
+## Build generated
+build/
+DerivedData
+
+## Various settings
+*.pbxuser
+!default.pbxuser
+*.mode1v3
+!default.mode1v3
+*.mode2v3
+!default.mode2v3
+*.perspectivev3
+!default.perspectivev3
+xcuserdata
+
+## Other
+*.xccheckout
+*.moved-aside
+*.xcuserstate
+*.xcscmblueprint
+
+## Obj-C/Swift specific
+*.hmap
+*.ipa
+
+# CocoaPods
+#
+# We recommend against adding the Pods directory to your .gitignore. However
+# you should judge for yourself, the pros and cons are mentioned at:
+# http://guides.cocoapods.org/using/using-cocoapods.html#should-i-check-the-pods-directory-into-source-control
+#
+Pods/
+
+# SwiftLint
+
+SwiftLint.pkg
+SwiftLintFramework.framework.zip
+benchmark_*
+portable_swiftlint.zip
+swiftlint_linux.zip
+osscheck/
+docs/
+rule_docs/
+
+# Swift Package Manager
+#
+# Add this line if you want to avoid checking in source code from Swift Package Manager dependencies.
+Packages/
+.build/
+.swiftpm/
+
+# macOS
+.DS_Store
+
+# Bundler
+.bundle/
+bundle/
