+++ Pampaty/AppDefault/AppDelegate.swift
@@ -34,6 +34,4 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
     // Use this method to release any resources that were specific to the discarded scenes, as they will not return.
   }
 
-
 }
-
