+++ Pampaty/AppDefault/SceneDelegate.swift
@@ -8,22 +8,21 @@
 import UIKit
 
 class SceneDelegate: UIResponder, UIWindowSceneDelegate {
-  
-  var window: UIWindow?
-  static let shared = SceneDelegate()
-  let navigationController = RootNavigation()
-  
-  func scene(_ scene: UIScene, willConnectTo session: UISceneSession, options connectionOptions: UIScene.ConnectionOptions) {
-    guard let scene = (scene as? UIWindowScene) else { return }
-    
-    let window = UIWindow(windowScene: scene)
-    let startViewController = GameSettingsVC()
-    navigationController.viewControllers.append(startViewController)
-    
-    window.rootViewController = navigationController
-    self.window = window
-    window.makeKeyAndVisible()
-  }
-  
-}
 
+	var window: UIWindow?
+	static let shared = SceneDelegate()
+	let navigationController = RootNavigation()
+
+	func scene(_ scene: UIScene, willConnectTo session: UISceneSession, options connectionOptions: UIScene.ConnectionOptions) {
+		guard let scene = (scene as? UIWindowScene) else { return }
+
+		let window = UIWindow(windowScene: scene)
+		let startViewController = GameSettingsVC()
+		navigationController.viewControllers.append(startViewController)
+
+		window.rootViewController = navigationController
+		self.window = window
+		window.makeKeyAndVisible()
+	}
+
+}
