+++ Pampaty/Components/DefaultTableView/CellViewModelProtocol.swift
@@ -1,17 +1,15 @@
 import UIKit
 
-protocol CellProtocol {
+protocol CellProtocol: AnyObject {
   func onValueChanged()
 }
 
-protocol CellViewModelProtocol {
+protocol CellViewModelProtocol: AnyObject {
   var title: String? { get set }
   var subtitle: String? { get set }
   var accessory: UIView? { get set }
   var accessoryDescription: String? { get set }
   var selectionStyle: UITableViewCell.SelectionStyle { get set }
   func onCellSelection()
-  var delegate: CellProtocol? { get set }
+	var delegate: CellProtocol? { get set }
 }
-
-
