+++ Pampaty/Components/DefaultTableView/TableViewItems.swift
@@ -4,4 +4,3 @@ struct TableSection {
   var title: String
   var items: [CellViewModelProtocol]
 }
-
