+++ Pampaty/Screens/GameSettings/GameSettingsVM.swift
@@ -1,20 +1,20 @@
 import UIKit
 
 class GameSettingsVM {
-  var tableSections: [TableSection] = [
-    
-    TableSection(title: "Команды", items: [
-      NumberOfTeamsCellVM(),
-      TeamsInGameCellVM()
-    ]),
-    
-    TableSection(title: "Игра", items: [
-      RoundTimingCellVM(),
-      RoundNumbersCellVM()
-    ]),
-    
-    TableSection(title: "Фишки", items: [
-      ModificationCellVM()
-    ])
-  ]
+	var tableSections: [TableSection] = [
+
+		TableSection(title: "Команды", items: [
+			NumberOfTeamsCellVM(),
+			TeamsInGameCellVM()
+		]),
+
+		TableSection(title: "Игра", items: [
+			RoundTimingCellVM(),
+			RoundNumbersCellVM()
+		]),
+
+		TableSection(title: "Фишки", items: [
+			ModificationCellVM()
+		])
+	]
 }
