--- Pampaty/Screens/GameSettings/SettingsTableItems.swift
@@ -1,1 +1,0 @@
-
