+++ Pampaty/Screens/Navigation/RootNavigation.swift
@@ -1,12 +1,12 @@
 import UIKit
 
 class RootNavigation: UINavigationController {
-  
+
   override func viewDidLoad() {
     super.viewDidLoad()
     setup()
   }
-  
+
   private func setup() {}
-  
+
 }
