+++ Pampaty/Screens/Teams/CellsVMs/TeamsInGameCellVM.swift
@@ -4,18 +4,17 @@ class TeamsInGameCellVM: CellViewModelProtocol {
   func onCellSelection() {
     AppDelegate.sceneDelegate?.navigationController.pushViewController(TeamsVC(), animated: true)
   }
-  
+
   var title: String? = "Команд в игре"
-  
+
   var subtitle: String?
-  
+
   var accessory: UIView? = UIImageView(image: UIImage(systemName: "chevron.right")?.withRenderingMode(.alwaysTemplate))
-  
+
   var accessoryDescription: String? = "Изменить"
-  
+
   var selectionStyle: UITableViewCell.SelectionStyle = .none
-  
-  var delegate: CellProtocol?
-  
-  
+
+  weak var delegate: CellProtocol?
+
 }
