+++ Pampaty/Screens/Teams/TeamsVM.swift
@@ -1,9 +1,9 @@
 import Foundation
 
 class TeamsVM {
-  
-  var items: [CellViewModelProtocol] = []
-  
-  init() { }
-  
+
+	var items: [CellViewModelProtocol] = []
+
+	init() { }
+
 }
