+++ PampatyTests/Screens/Settings/SettingsCellViewModelNumberOfTeamsTests.swift
@@ -2,11 +2,7 @@ import XCTest
 @testable import Pampaty
 
 class SettingsCellViewModelNumberOfTeamsTests: XCTestCase {
-  
+
   var sut = NumberOfTeamsCellVM()
-  
-  
-  
-  
-   
+
 }
