--- git.gitignore
@@ -1,99 +1,0 @@
-{\rtf1\ansi\ansicpg1251\cocoartf2578
-\cocoatextscaling0\cocoaplatform0{\fonttbl\f0\fmodern\fcharset0 Courier;}
-{\colortbl;\red255\green255\blue255;\red0\green0\blue0;}
-{\*\expandedcolortbl;;\cssrgb\c0\c0\c0;}
-\paperw11900\paperh16840\margl1440\margr1440\vieww11520\viewh8400\viewkind0
-\deftab720
-\pard\pardeftab720\partightenfactor0
-
-\f0\fs24 \cf2 \expnd0\expndtw0\kerning0
-\outl0\strokewidth0 \strokec2 \
-# Created by https://www.toptal.com/developers/gitignore/api/swift\
-# Edit at https://www.toptal.com/developers/gitignore?templates=swift\
-\
-### Swift ###\
-# Xcode\
-#\
-# gitignore contributors: remember to update Global/Xcode.gitignore, Objective-C.gitignore & Swift.gitignore\
-\
-## User settings\
-*.idea/workspace.xml
-\
-## compatibility with Xcode 8 and earlier (ignoring not required starting Xcode 9)\
-*.xcscmblueprint\
-*.xccheckout\
-\
-## compatibility with Xcode 3 and earlier (ignoring not required starting Xcode 4)\
-build/\
-DerivedData/\
-*.moved-aside\
-*.pbxuser\
-!default.pbxuser\
-*.mode1v3\
-!default.mode1v3\
-*.mode2v3\
-!default.mode2v3\
-*.perspectivev3\
-!default.perspectivev3\
-\
-## Obj-C/Swift specific\
-*.hmap\
-\
-## App packaging\
-*.ipa\
-*.dSYM.zip\
-*.dSYM\
-\
-## Playgrounds\
-timeline.xctimeline\
-playground.xcworkspace\
-\
-# Swift Package Manager\
-# Add this line if you want to avoid checking in source code from Swift Package Manager dependencies.\
-# Packages/\
-# Package.pins\
-# Package.resolved\
-# *.xcodeproj\
-# Xcode automatically generates this directory with a .xcworkspacedata file and xcuserdata\
-# hence it is not needed unless you have added a package configuration file to your project\
-# .swiftpm\
-\
-.build/\
-\
-# CocoaPods\
-# We recommend against adding the Pods directory to your .gitignore. However\
-# you should judge for yourself, the pros and cons are mentioned at:\
-# https://guides.cocoapods.org/using/using-cocoapods.html#should-i-check-the-pods-directory-into-source-control\
-# Pods/\
-# Add this line if you want to avoid checking in source code from the Xcode workspace\
-# *.xcworkspace\
-\
-# Carthage\
-# Add this line if you want to avoid checking in source code from Carthage dependencies.\
-# Carthage/Checkouts\
-\
-Carthage/Build/\
-\
-# Accio dependency management\
-Dependencies/\
-.accio/\
-\
-# fastlane\
-# It is recommended to not store the screenshots in the git repo.\
-# Instead, use fastlane to re-generate the screenshots whenever they are needed.\
-# For more information about the recommended setup visit:\
-# https://docs.fastlane.tools/best-practices/source-control/#source-control\
-\
-fastlane/report.xml\
-fastlane/Preview.html\
-fastlane/screenshots/**/*.png\
-fastlane/test_output\
-\
-# Code Injection\
-# After new code Injection tools there's a generated folder /iOSInjectionProject\
-# https://github.com/johnno1962/injectionforxcode\
-\
-iOSInjectionProject/\
-\
-# End of https://www.toptal.com/developers/gitignore/api/swift\
-}
